Compile parts of test-util.h that we can once, link with unit tests
