+++ cpp/src/arrow/dbi/hiveserver2/CMakeLists.txt
@@ -105,9 +105,9 @@ set_property(SOURCE ${ARROW_HIVESERVER2_SRCS}
   " -Wno-shadow-field")
 
 set(ARROW_HIVESERVER2_TEST_LINK_LIBS
-  ${ARROW_TEST_LINK_LIBS}
   arrow_hiveserver2_static
   arrow_hiveserver2_thrift
+  ${ARROW_TEST_LINK_LIBS}
   thriftstatic)
 
 if (ARROW_BUILD_TESTS)
